decompose server class
